Move database creation into helper class.
Move database related flags and the code to create one into a new helper
class and remove the duplicated code from the servers.
