+++ cpp/server/xjson-server.cc
@@ -31,6 +31,7 @@
 #include "monitoring/registry.h"
 #include "server/metrics.h"
 #include "server/server.h"
+#include "server/server_helper.h"
 #include "server/x_json_handler.h"
 #include "util/etcd.h"
 #include "util/fake_etcd.h"
@@ -44,7 +45,6 @@
 DEFINE_string(server, "localhost", "Server host");
 DEFINE_int32(port, 9999, "Server port");
 DEFINE_string(key, "", "PEM-encoded server private key file");
-DEFINE_string(leveldb_db, "", "LevelDB database for leaf and tree storage");
 DEFINE_int32(log_stats_frequency_seconds, 3600,
              "Interval for logging summary statistics. Approximate: the "
              "server will log statistics if in the beginning of its select "
@@ -90,6 +90,7 @@ using cert_trans::LoggedEntry;
 using cert_trans::ReadPrivateKey;
 using cert_trans::ScopedLatency;
 using cert_trans::Server;
+using cert_trans::ServerHelper;
 using cert_trans::SplitHosts;
 using cert_trans::ThreadPool;
 using cert_trans::TreeSigner;
@@ -296,11 +297,8 @@ int main(int argc, char* argv[]) {
   CHECK_EQ(pkey.status(), util::Status::OK);
   LogSigner log_signer(pkey.ValueOrDie());
 
-  if (FLAGS_leveldb_db.empty()) {
-    std::cerr << "Must specify database.";
-    exit(1);
-  }
-  LevelDB db(FLAGS_leveldb_db);
+  ServerHelper::EnsureValidatorsRegistered();
+  Database* db = ServerHelper::ProvideDatabase();
 
   shared_ptr<libevent::Base> event_base(make_shared<libevent::Base>());
   ThreadPool internal_pool(8);
@@ -329,13 +327,13 @@ int main(int argc, char* argv[]) {
 
   ThreadPool http_pool(FLAGS_num_http_server_threads);
 
-  Server server(options, event_base, &internal_pool, &http_pool, &db,
+  Server server(options, event_base, &internal_pool, &http_pool, db,
                 etcd_client.get(), &url_fetcher, &log_verifier);
   server.Initialise(false /* is_mirror */);
 
   Frontend frontend(
-      new FrontendSigner(&db, server.consistent_store(), &log_signer));
-  XJsonHttpHandler handler(server.log_lookup(), &db,
+      new FrontendSigner(db, server.consistent_store(), &log_signer));
+  XJsonHttpHandler handler(server.log_lookup(), db,
                            server.cluster_state_controller(), &frontend,
                            &internal_pool, event_base.get());
 
@@ -344,7 +342,7 @@ int main(int argc, char* argv[]) {
   handler.Add(server.http_server());
 
   TreeSigner<LoggedEntry> tree_signer(
-      std::chrono::duration<double>(FLAGS_guard_window_seconds), &db,
+      std::chrono::duration<double>(FLAGS_guard_window_seconds), db,
       server.log_lookup()->GetCompactMerkleTree(new Sha256Hasher),
       server.consistent_store(), &log_signer);
